pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"


[project]
name = "icd10"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=4.5.0",
    "dotenv>=0.9.9",
    "faiss-cpu>=1.13.2",
    "google-genai>=1.61.0",
    "ipykernel>=7.1.0",
    "mlflow==3.9.0",
    "nest-asyncio>=1.6.0",
    "openai>=2.16.0",
    "pandas<3.0.0",
    "polars>=1.37.1",
    "protobuf<5.0.0",
    "psycopg>=3.3.2",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "sentencepiece>=0.2.1",
    "streamlit>=1.54.0",
    "tenacity>=9.1.2",
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    "torchvision>=0.25.0",
    "tqdm>=4.67.3",
    "transformers>=5.0.0",
]


[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
include = ["icd_codes*"]
